+++ frontend/src/app/shared/components/grids/widgets/news-beta/news-beta.component.ts
@@ -35,6 +35,6 @@ import { AbstractTurboWidgetComponent } from 'core-app/shared/components/grids/w
   standalone: false,
 })
 export class WidgetNewsBetaComponent extends AbstractTurboWidgetComponent {
-  override frameId = 'overviews-widgets-news-component';
+  override frameId = 'grids-widgets-news-component';
   override name = 'news';
 }
