+++ modules/grids/app/components/grids/widgets/news_component.html.erb
