+++ modules/grids/lib/grids/engine.rb
@@ -1,9 +1,21 @@
 module Grids
   class Engine < ::Rails::Engine
+    engine_name :grids
+
     include OpenProject::Plugins::ActsAsOpEngine
 
     add_api_path :attachments_by_grid do |id|
       "#{root}/grids/#{id}/attachments"
+    end
+
+    initializer "grids.permissions" do
+      Rails.application.reloader.to_prepare do
+        OpenProject::AccessControl.permission(:view_news)
+          .controller_actions
+          .push(
+            "grids/widgets/news/show"
+          )
+      end
     end
 
     config.to_prepare do
