+++ modules/grids/spec/routing/widget_routing_spec.rb
@@ -28,9 +28,42 @@
 # See COPYRIGHT and LICENSE files for more details.
 #++
 
-class Overviews::WidgetController < Grids::BaseInProjectController
-  include OpTurbo::ComponentStream
-  include OpTurbo::FlashStreamHelper
+require "rails_helper"
 
-  layout -> { "turbo_rails/frame" }
+RSpec.describe Grids::WidgetController do
+  describe "news routing" do
+    describe "GET #show" do
+      context "for root" do
+        it do
+          expect(get("/widgets/news"))
+            .to route_to(controller: "grids/widgets/news", action: "show")
+        end
+      end
+
+      context "with project" do
+        it do
+          expect(get("/projects/my-project/widgets/news"))
+            .to route_to(controller: "grids/widgets/news", action: "show", project_id: "my-project")
+        end
+      end
+    end
+  end
+
+  describe "news named routing" do
+    describe "GET #show" do
+      context "for root" do
+        it do
+          expect(get(widgets_news_path))
+            .to route_to(controller: "grids/widgets/news", action: "show")
+        end
+      end
+
+      context "with project" do
+        it do
+          expect(get(project_widgets_news_path("my-project")))
+            .to route_to(controller: "grids/widgets/news", action: "show", project_id: "my-project")
+        end
+      end
+    end
+  end
 end
