+++ modules/overviews/config/routes.rb
@@ -17,15 +17,8 @@ Rails.application.routes.draw do
 
         namespace :widgets do
           resource :project_status, only: %i[show update]
-          resource :news, only: %i[show]
         end
       end
-    end
-  end
-
-  scope module: "overviews" do
-    namespace :widgets do
-      resource :news, only: %i[show]
     end
   end
